Allowed Agents to be defined in other directories.
Also:
 - Because we now can, move example bots to their own directories.
 - Refactor grpc_demo_agent to keep non-usable code out of grpcsupport.
