+++ agents/grpc_demo_agent/grpc_demo_agent.cfg
@@ -1,6 +1,6 @@
 [Bot Location]
 # Path to module from runner
-agent_module = grpcsupport.grpc_demo_agent
+agent_module = grpc_demo_agent
 
 [Participant Loadout]
 # Name that will be displayed in game
