+++ agents/grpc_demo_agent/grpc_demo_agent.py
@@ -1,0 +1,3 @@
+from grpcsupport.grpc_client import make_grpc_agent
+
+Agent = make_grpc_agent('localhost:34865')
