+++ rlbot.cfg
@@ -5,16 +5,17 @@ num_participants = 2
 [Participant Configuration]
 # Put the name of your bot config file here.  Only total_num_participants config files will be read!
 # Everything needs a config, even players and default bots.  We still set loadouts and names from config!
-participant_config_0 = atba.cfg
-participant_config_1 = atba.cfg
-participant_config_2 = atba.cfg
-participant_config_3 = atba.cfg
-participant_config_4 = atba.cfg
-participant_config_5 = atba.cfg
-participant_config_6 = atba.cfg
-participant_config_7 = atba.cfg
-participant_config_8 = atba.cfg
-participant_config_9 = atba.cfg
+participant_config_0 = ./agents/atba/atba.cfg
+# e.g. Could use       ./agents/grpc_demo_agent/grpc_demo_agent.cfg
+participant_config_1 = ./agents/atba/atba.cfg
+participant_config_2 = ./agents/atba/atba.cfg
+participant_config_3 = ./agents/atba/atba.cfg
+participant_config_4 = ./agents/atba/atba.cfg
+participant_config_5 = ./agents/atba/atba.cfg
+participant_config_6 = ./agents/atba/atba.cfg
+participant_config_7 = ./agents/atba/atba.cfg
+participant_config_8 = ./agents/atba/atba.cfg
+participant_config_9 = ./agents/atba/atba.cfg
 
 # team 0 shoots on positive goal, team 1 shoots on negative goal
 participant_team_0 = 0
